requirements.txt
# Requirements for Chess vs Stockfish
pygame>=2.6.0
python-chess>=1.999
